% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{colourSets}
\alias{colourSets}
\alias{nice_colours}
\alias{pastel_colours}
\alias{rich_colours}
\alias{bright_colours_transparent}
\alias{bright_colours_opaque}
\title{gmoviz colour sets}
\format{Character vectors of 34 hex colours.}
\source{
Many of the colours are from, or inspired by ColorBrewer
\url{http://colorbrewer2.org/}.
}
\usage{
nice_colours

pastel_colours

rich_colours

bright_colours_transparent

bright_colours_opaque
}
\description{
gmoviz colour sets

\code{nice_colours}: The default colour set. Medium brightness, light
colours. Designed for use on a white/light coloured background.

\code{pastel_colours}:A set of pale/pastel colours, modelled on the
\code{nice_colours} set but less saturated. Designed for use on a
white/light background.

\code{rich_colours}: A set of bright, vibrant colours (but not neon, like
the bright_colours_transparent). Designed for use on any sort of background.

\code{bright_colours_transparent}: A set of very bright (neon) colours with
slight transparency. Designed for use on a black background.

\code{bright_colours_opaque}: A set of very bright (neon) colours without
transparency. Designed for use on a black background.
}
\details{
Due to the often high number of sectors being plotted with gmoviz
(e.g. 20+ when plotting each chromosome), a number of 'colour sets' have
been included for convenience.
}
\keyword{datasets}
